requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
fpdf2>=2.7.0
gspread>=5.12.0
google-auth>=2.25.0
